requirements.txt
docker-compose==1.26.0
Django==2.2.13
djangorestframework>=3.11.0,<3.12.0
django-nested-admin==3.3.2